Variety of small fixes
