+++ README.md
@@ -32,6 +32,8 @@ NOTE: if this takes too long, it is possible to run the Axiom Profiler with a pr
 
 Similarly, if you have a log file which takes too long to load into the Axiom Profiler, hitting Cancel will cause the tool to work with the portion loaded so far.
 
+To correctly parse the log file, we impose a few [restrictions](smt-log-parser/design-docs/restrictions.md) on the smt2 file given to z3.
+
 ## Obtaining Z3 logs from various verification tools that use Z3 (feel free to add more)
 
 ### Boogie
