+++ smt-log-parser/src/parsers/z3/mod.rs
@@ -6,14 +6,17 @@ use crate::{Error, FResult, Result};
 pub mod egraph;
 pub mod inst;
 pub mod inter_line;
+/// Original Z3 log parser. Works with Z3 v.4.12.1, should work with other versions
+/// as long as the log format is the same for the important line cases.
+/// Compare with the log files in the `logs/` folder to see if this is the case.
+pub mod parse;
 pub mod stack;
 pub mod stm2;
 pub mod synthetic;
 pub mod terms;
-/// Original Z3 log parser. Works with Z3 v.4.12.1, should work with other versions
-/// as long as the log format is the same for the important line cases.
-/// Compare with the log files in the `logs/` folder to see if this is the case.
-pub mod z3parser;
+mod z3parser;
+
+pub use z3parser::*;
 
 impl<T: Z3LogParser + LogParserHelper> LogParser for T {
     fn is_line_start(&mut self, first_byte: u8) -> bool {
