+++ smt-log-parser/tests/data/ast_size.json
@@ -1,0 +1,1 @@
+[{"Ok":1},{"Ok":1},{"Ok":1},{"Ok":1},{"Ok":1},{"Ok":1},{"Err":6},{"Err":7},{"Err":6},{"Err":6},{"Err":6},{"Err":6},{"Err":6},{"Err":6},{"Err":7},{"Err":7},{"Ok":1},{"Err":17},{"Err":18},{"Err":19},{"Err":20},{"Err":21},{"Err":22},{"Err":23},{"Err":24},{"Err":25},{"Err":26},{"Err":27},{"Err":28},{"Err":29},{"Err":30},{"Err":31},{"Err":25},{"Ok":1},{"Err":34},{"Err":35},{"Err":36},{"Err":37},{"Err":38},{"Err":39},{"Err":40},{"Err":41},{"Err":42},{"Err":43},{"Err":44},{"Err":45},{"Err":46},{"Err":47},{"Err":48},{"Err":42},{"Ok":1},{"Err":51},{"Err":52},{"Err":51},{"Err":51},{"Err":52},{"Err":51},{"Err":52},{"Err":52},{"Err":51},{"Ok":1},{"Err":52},{"Err":51},{"Err":51},{"Err":51},{"Err":51},{"Err":52},{"Err":52},{"Err":51},{"Err":51},{"Err":51},{"Err":51},{"Err":51},{"Ok":1},{"Err":51},{"Err":75},{"Ok":1},{"Err":75},{"Err":75},{"Ok":1},{"Err":52},{"Ok":1},{"Ok":1},{"Err":75},{"Err":75},{"Err":52},{"Err":52},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Ok":1},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Ok":1},{"Err":52},{"Err":52},{"Err":75},{"Err":52},{"Err":52},{"Err":75},{"Ok":3},{"Err":52},{"Err":52},{"Err":52},{"Err":52},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Ok":1},{"Err":75},{"Ok":1},{"Ok":3},{"Ok":1},{"Ok":3},{"Ok":5},{"Ok":9},{"Ok":3},{"Ok":7},{"Ok":5},{"Ok":11},{"Ok":5},{"Ok":6},{"Ok":12},{"Ok":12},{"Ok":3},{"Ok":5},{"Ok":7},{"Ok":13},{"Ok":8},{"Ok":15},{"Ok":14},{"Ok":12},{"Ok":22},{"Ok":3},{"Ok":5},{"Ok":4},{"Ok":6},{"Ok":12},{"Ok":13},{"Ok":12},{"Ok":13},{"Ok":26},{"Ok":7},{"Ok":14},{"Ok":8},{"Ok":14},{"Ok":16},{"Ok":29},{"Ok":17},{"Ok":31},{"Ok":30},{"Ok":18},{"Ok":32},{"Ok":35},{"Ok":30},{"Err":52},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Ok":1},{"Err":75},{"Ok":7},{"Ok":14},{"Ok":28},{"Ok":27},{"Ok":26},{"Err":52},{"Err":51},{"Err":75},{"Err":75},{"Ok":7},{"Ok":13},{"Ok":25},{"Err":52},{"Err":51},{"Err":75},{"Ok":7},{"Ok":13},{"Ok":25},{"Err":52},{"Err":51},{"Err":75},{"Ok":7},{"Ok":13},{"Ok":25},{"Err":52},{"Err":51},{"Err":51},{"Err":75},{"Ok":7},{"Ok":11},{"Err":75},{"Err":51},{"Err":52},{"Err":52},{"Err":52},{"Err":52},{"Ok":1},{"Err":52},{"Err":52},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Ok":1},{"Err":75},{"Ok":7},{"Ok":14},{"Ok":16},{"Ok":29},{"Err":52},{"Err":52},{"Err":75},{"Err":52},{"Err":75},{"Err":52},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Ok":1},{"Err":75},{"Ok":17},{"Ok":14},{"Ok":29},{"Ok":16},{"Ok":33},{"Ok":33},{"Ok":33},{"Err":52},{"Err":52},{"Err":51},{"Err":75},{"Err":75},{"Ok":5},{"Ok":17},{"Ok":9},{"Ok":33},{"Ok":18},{"Ok":33},{"Ok":18},{"Ok":17},{"Ok":9},{"Ok":5},{"Err":75},{"Err":75},{"Err":51},{"Err":52},{"Err":52},{"Err":52},{"Ok":1},{"Err":52},{"Err":52},{"Err":51},{"Ok":1},{"Err":51},{"Err":75},{"Ok":1},{"Err":75},{"Err":75},{"Ok":7},{"Ok":17},{"Ok":17},{"Ok":33},{"Ok":33},{"Err":52},{"Err":75},{"Ok":17},{"Ok":33},{"Err":52},{"Err":75},{"Ok":17},{"Ok":33},{"Err":52},{"Err":75},{"Ok":17},{"Ok":33},{"Err":52},{"Err":75},{"Ok":17},{"Ok":33},{"Err":52},{"Err":75},{"Ok":17},{"Ok":33},{"Err":52},{"Err":51},{"Err":75},{"Ok":5},{"Ok":9},{"Ok":18},{"Ok":5},{"Ok":7},{"Ok":9},{"Ok":11},{"Ok":23},{"Err":52},{"Err":52},{"Ok":11},{"Ok":11},{"Ok":3},{"Ok":5},{"Ok":10},{"Ok":21},{"Err":51},{"Err":51},{"Ok":4},{"Ok":5},{"Ok":7},{"Ok":5},{"Ok":7},{"Ok":13},{"Ok":15},{"Ok":34},{"Ok":7},{"Ok":3},{"Ok":11},{"Ok":5},{"Ok":9},{"Ok":5},{"Ok":11},{"Ok":7},{"Ok":15},{"Ok":7},{"Ok":15},{"Ok":15},{"Ok":15},{"Ok":19},{"Ok":17},{"Ok":14},{"Ok":49},{"Ok":16},{"Ok":36},{"Ok":7},{"Ok":9},{"Ok":13},{"Ok":15},{"Ok":12},{"Ok":26},{"Ok":14},{"Ok":30},{"Ok":11},{"Ok":13},{"Ok":28},{"Ok":29},{"Err":52},{"Err":52},{"Err":52},{"Err":52},{"Ok":13},{"Ok":13},{"Ok":5},{"Ok":6},{"Ok":5},{"Ok":7},{"Ok":9},{"Ok":7},{"Ok":15},{"Ok":17},{"Ok":40},{"Ok":3},{"Ok":9},{"Ok":4},{"Ok":11},{"Ok":3},{"Ok":7},{"Ok":13},{"Ok":5},{"Ok":5},{"Ok":13},{"Ok":11},{"Ok":7},{"Ok":15},{"Ok":9},{"Ok":9},{"Ok":3},{"Ok":13},{"Ok":5},{"Ok":11},{"Ok":15},{"Ok":31},{"Ok":17},{"Ok":35},{"Ok":9},{"Ok":15},{"Ok":17},{"Ok":35},{"Ok":35},{"Ok":30},{"Ok":71},{"Ok":32},{"Ok":42},{"Ok":9},{"Ok":11},{"Ok":7},{"Ok":9},{"Ok":17},{"Ok":19},{"Ok":46},{"Ok":9},{"Ok":19},{"Ok":11},{"Ok":11},{"Ok":15},{"Ok":13},{"Ok":6},{"Ok":14},{"Ok":8},{"Ok":18},{"Ok":7},{"Ok":16},{"Ok":17},{"Ok":19},{"Ok":23},{"Ok":21},{"Ok":14},{"Ok":61},{"Ok":16},{"Ok":48},{"Ok":7},{"Ok":9},{"Ok":17},{"Ok":40},{"Ok":9},{"Ok":3},{"Ok":13},{"Ok":5},{"Ok":11},{"Ok":30},{"Ok":71},{"Ok":29},{"Ok":60},{"Ok":70},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Err":75},{"Ok":14},{"Ok":16},{"Ok":18},{"Ok":18},{"Ok":1},{"Ok":1},{"Ok":1},{"Ok":1},{"Ok":1},{"Ok":1},{"Ok":1},{"Ok":1},{"Ok":1},{"Ok":1}]
